+++ typhon-engine/src/hydra/main.rs
@@ -269,7 +269,6 @@ async fn main() -> anyhow::Result<()> {
 
     let state = api::AppState {
         imports: Default::default(),
-        facts_cache: Default::default(),
         config: Arc::new(std::sync::RwLock::new(Arc::new(config))),
         engines: engine_host,
         store: shared_store.clone(),